requirements.txt
# Core dependencies for face recognition
face_recognition>=1.3.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Google Drive API
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Progress bar and utilities
tqdm>=4.66.0
